Moved formatcore/* to format/formatcore/*
Updated imports and added README.md in format/formatcore.
