+++ format/devops.go
@@ -7,7 +7,7 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/princjef/gomarkdoc/formatcore"
+	"github.com/princjef/gomarkdoc/format/formatcore"
 	"github.com/princjef/gomarkdoc/lang"
 )
