+++ format/formatcore/base.go
